Move spacer to `ui` package
